requirements.txt
ibapi==10.19.2
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
loguru==0.7.2
schedule==1.2.2
